requirement.txt
# Core Gemini / Google Gen AI SDK
google-genai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

